feat: null for params without values and undefined for missing params
BREAKING CHANGE: Query params now treat null and undefined differently. If there is a parameter in the URL but no value - its treated as `null`; however, if the paramter is completely missing from the URL - then its `undefined`. BEWARE: because of this, the `defaultValue` will be used exclusively if the value is `undefined` (since `null` is treated as a set value).